refactor filepaths; delegate logic
