+++ data/agentstats
@@ -1,3 +1,3 @@
-win_percentage=0.45
-loss_count=63
-win_count=51
+win_percentage=0.52
+win_count=106
+loss_count=96
